+++ lib/csharp/src/Transports.cs
@@ -237,7 +237,7 @@ namespace Agnos.Transports
                 AssertBeganWrite();
                 if (buffer.Length > 0)
                 {
-                    Packers.Int32.pack(buffer.Length, outputStream);
+                    Packers.Int32.pack((int)buffer.Length, outputStream);
                     Packers.Int32.pack(wseq, outputStream);
                     buffer.WriteTo(outputStream);
                     outputStream.Flush();
